use fixed bit size for sc_ulimb_t

+++ src/utils/arith/sc_mp.h
@@ -50,8 +50,8 @@
 #define SC_LIMB_B4(x)      ((x) << 16)
 typedef UINT128 sc_ulimb_big_t;
 typedef SINT128 sc_slimb_big_t;
-typedef unsigned long sc_ulimb_t;
-typedef signed long sc_slimb_t;
+typedef UINT64 sc_ulimb_t;
+typedef INT64 sc_slimb_t;
 typedef UINT32 sc_ulimb_half_t;
 typedef SINT32 sc_slimb_half_t;
 
@@ -68,8 +68,8 @@ typedef SINT32 sc_slimb_half_t;
 #define SC_LIMB_B4(x)      ((x) << 8)
 typedef UINT64 sc_ulimb_big_t;
 typedef SINT64 sc_slimb_big_t;
-typedef unsigned long sc_ulimb_t;
-typedef signed long sc_slimb_t;
+typedef UINT32 sc_ulimb_t;
+typedef INT32 sc_slimb_t;
 typedef UINT16 sc_ulimb_half_t;
 typedef SINT16 sc_slimb_half_t;
